requirements.txt
google-cloud-vision==0.39.0
requests==2.11.0
